+++ ngraph/src/main/java/org/bytedeco/ngraph/presets/ngraph.java
@@ -43,6 +43,7 @@ import java.lang.annotation.Target;
         "ngraph/frontend/onnxifi/backend_manager.hpp",
         "ngraph/descriptor/tensor.hpp",
         "ngraph/pass/pass_config.hpp",
+        "ngraph/runtime/executable.hpp",
         "ngraph/runtime/tensor.hpp",
         "ngraph/runtime/backend.hpp",
         "ngraph/runtime/cpu/cpu_backend.hpp",
@@ -72,6 +73,7 @@ import java.lang.annotation.Target;
         "ngraph/dimension.hpp",
         "ngraph/rank.hpp",
         "ngraph/partial_shape.hpp",
+        "ngraph/check.hpp",
         "ngraph/node.hpp",
         "ngraph/frontend/onnxifi/onnxifi.h",
 //        "core/node.hpp",
@@ -112,14 +114,16 @@ public class ngraph implements InfoMapper {
                .put(new Info("ngraph::element::from<ngraph::bfloat16>").javaNames("fromNGraphBFloat16"))
 
                .put(new Info("ngraph::op::util::BinaryElementwiseArithmetic", "ngraph::op::ScalarConstantLike").purify(true))
-               .put(new Info("std::shared_ptr<ngraph::op::Result>","std::shared_ptr<op::Result>").annotations("@SharedPtr").pointerTypes("Result"))
+               .put(new Info("std::shared_ptr<ngraph::op::Result>").annotations("@SharedPtr").pointerTypes("Result"))
                .put(new Info("std::shared_ptr<ngraph::runtime::Tensor>").annotations("@SharedPtr").pointerTypes("Tensor"))
 //               .put(new Info("ngraph::Node").purify(false).virtualize())
-               .put(new Info("std::shared_ptr<const Node>", "std::shared_ptr<ngraph::Node>").annotations("@SharedPtr").pointerTypes("Node"))
-               .put(new Info("std::shared_ptr<ngraph::op::Parameter>", "std::shared_ptr<op::Parameter>").annotations("@SharedPtr").pointerTypes("Parameter"))
-               .put(new Info("std::shared_ptr<ngraph::Function>", "std::shared_ptr<Function>").annotations("@SharedPtr").pointerTypes("Function"))
-               .put(new Info("std::shared_ptr<const ngraph::Function>", "std::shared_ptr<const Function>").annotations("@Cast(\"const ngraph::Function*\") @SharedPtr").pointerTypes("Function"))
-               .put(new Info("std::enable_shared_from_this<ngraph::Node>", "std::enable_shared_from_this<Node>", "std::enable_shared_from_this<ngraph::runtime::cpu::CPU_ExternalFunction>").pointerTypes("Pointer"))
+               .put(new Info("std::shared_ptr<ngraph::Node>").annotations("@SharedPtr").pointerTypes("Node"))
+               .put(new Info("std::shared_ptr<const ngraph::Node>").annotations("@Cast(\"const ngraph::Node*\") @SharedPtr").pointerTypes("Node"))
+               .put(new Info("std::shared_ptr<ngraph::op::Constant>").annotations("@SharedPtr").pointerTypes("Constant"))
+               .put(new Info("std::shared_ptr<ngraph::op::Parameter>").annotations("@SharedPtr").pointerTypes("Parameter"))
+               .put(new Info("std::shared_ptr<ngraph::Function>").annotations("@SharedPtr").pointerTypes("Function"))
+               .put(new Info("std::shared_ptr<const ngraph::Function>").annotations("@Cast(\"const ngraph::Function*\") @SharedPtr").pointerTypes("Function"))
+               .put(new Info("std::enable_shared_from_this<ngraph::Node>", "std::enable_shared_from_this<ngraph::runtime::cpu::CPU_ExternalFunction>").pointerTypes("Pointer"))
                .put(new Info("std::runtime_error").cast().pointerTypes("Pointer"))
                .put(new Info("std::list<std::shared_ptr<Node> >", "std::pair<std::shared_ptr<ngraph::op::Result>,std::shared_ptr<ngraph::op::Parameter> >", "std::deque<ngraph::Node::descriptor::Input>", "std::deque<descriptor::Output>", "std::set<ngraph::Node::descriptor::Input*>", "std::unordered_set<descriptor::Tensor*>", "std::stringstream", "ngraph::Node::has_same_type", "ngraph::descriptor::Tensor::set_tensor_layout", "ngraph::runtime::cpu::CPU_ExternalFunction::get_executor", "ngraph::runtime::cpu::CPU_ExternalFunction::get_callees", "ngraph::runtime::cpu::CPU_ExternalFunction::get_halide_functions", "ngraph::runtime::cpu::CPU_ExternalFunction::get_subgraph_params", "ngraph::runtime::cpu::CPU_ExternalFunction::get_subgraph_param_sizes", "ngraph::runtime::cpu::CPU_ExternalFunction::get_subgraph_param_ptrs", "ngraph::runtime::cpu::CPU_ExternalFunction::get_parameter_layout_descriptors", "ngraph::runtime::cpu::CPU_ExternalFunction::get_result_layout_descriptors", "ngraph::runtime::cpu::CPU_ExternalFunction::get_mkldnn_emitter", "ngraph::runtime::cpu::CPU_ExternalFunction::add_state", "ngraph::runtime::cpu::CPU_ExternalFunction::add_state", "ngraph::runtime::cpu::CPU_ExternalFunction::get_functors", "ngraph::runtime::cpu::CPU_Backend::make_call_frame", "ngraph::onnxifi::BackendManager::unregister", "ngraph::onnxifi::BackendManager::get", "ngraph::onnx_import::register_operator").skip())
                .put(new Info("ONNXIFI_ABI", "ONNXIFI_PUBLIC", "ONNXIFI_CHECK_RESULT", "NGRAPH_API").cppTypes().annotations())
@@ -135,6 +139,7 @@ public class ngraph implements InfoMapper {
                .put(new Info("std::unordered_map<std::string,void*>").pointerTypes("StringVoidMap").define())
                .put(new Info("std::vector<std::shared_ptr<ngraph::op::Parameter> >", "std::vector<std::shared_ptr<op::Parameter> >").pointerTypes("NgraphParameterVector").define())
                .put(new Info("std::vector<std::shared_ptr<ngraph::Node> >").pointerTypes("NgraphNodeVector").define())
+               .put(new Info("std::vector<std::shared_ptr<ngraph::op::Constant> >").pointerTypes("NgraphOpConstantVector").define())
                .put(new Info("std::vector<std::shared_ptr<ngraph::runtime::Tensor> >", "std::vector<std::shared_ptr<runtime::Tensor> >").pointerTypes("NgraphTensorVector").define())
                .put(new Info("std::vector<std::shared_ptr<ngraph::Function> >").pointerTypes("NgraphFunctionVector").define());